chore: review float64 declaration for static numbers

+++ internal/app/flow/dto.go
@@ -67,7 +67,7 @@ func (r updateFlowInputDto) validate() error {
 		validation.Field(&r.Description, validation.NilOrNotEmpty),
 		validation.Field(&r.Active, validation.In(true, false)),
 		validation.Field(&r.Fallback, validation.In(true, false)),
-		validation.Field(&r.CurrentServePct, validation.Min(float64(0)), validation.Max(float64(100))),
+		validation.Field(&r.CurrentServePct, validation.Min(0.0), validation.Max(100.0)),
 	)
 }
